feat: Add UpdtUser function to update user information
This commit adds the UpdtUser function to the backend.ts file. The function allows for updating user information in the 'users' collection of the database. It takes an object containing the updated user information and the user ID as parameters. If successful, it returns the updated user object. If an error occurs, it returns the error object.